+++ apps/client/src/pages/auth/utils/apple-login.ts
